apps/basic/requirements.txt
fastapi
uvicorn[standard]
anthropic
python-dotenv
websockets 